refactor(desktop): separate renderer feature boundaries
