+++ apps/desktop/src/renderer/main.tsx
@@ -8,29 +8,37 @@ import {
   groupAuthorizationOperations,
   type ApplicationAccessGroup
 } from "@mdbase/connect-ui/access";
-import {
-  MDBASE_MARK_VIEW_BOX,
-  mdbaseMarkAccentRect,
-  mdbaseMarkInkRects
-} from "@mdbase/connect-ui/brand";
 import { applyThemePreference, loadThemePreference, saveThemePreference, type ThemePreference } from "@mdbase/connect-ui/theme";
 import "@mdbase/connect-ui/styles.css";
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { createRoot } from "react-dom/client";
-import { presentConnection, type ConnectionDotState } from "./connection-state.mjs";
+import { Collections } from "./collections-view";
+import { presentConnection } from "./connection-state.mjs";
+import {
+  Brand,
+  Empty,
+  NavButton,
+  PairingPanel,
+  SectionHeading,
+  SettingSwitch,
+  StatusDot
+} from "./ui-components";
+import {
+  allOperations,
+  hasContract,
+  host,
+  hostedCollectionCompatible,
+  message,
+  neededProvisions,
+  plural,
+  provisionNames,
+  relativeTime,
+  scopeDescription,
+  type AuthorizationCollection,
+  type Route
+} from "./view-model";
 import "./styles.css";
 
-type Route = "overview" | "collections" | "access" | "activity" | "settings";
-interface AuthorizationCollection {
-  id: string;
-  display_name: string;
-  spec_version: string;
-  contracts: ContractRequirement[];
-  kind: "local" | "hosted";
-  provisionable: boolean;
-}
-
-const allOperations = ["describe", "changes", "read", "query", "list_views", "execute_view", "read_view_source", "create", "update", "rename", "delete", "create_view_source", "update_view_source", "delete_view_source", "validate", "read_type", "create_type", "update_type", "list_timers", "put_timer", "cancel_timer", "reconcile_timers"];
 const routeCopy: Record<Route, { eyebrow: string; title: string; lede: string }> = {
   overview: {
     eyebrow: "This computer",
@@ -376,503 +384,6 @@ function Overview({ status, cloud, access, collectionCount, busy, onNavigate, on
 
 function OverviewRow({ label, value, detail, action, onClick }: { label: string; value: string; detail: string; action: string; onClick(): void }) {
   return <div className="overview-row"><span>{label}</span><div><strong>{value}</strong><small>{detail}</small></div><button className="quiet-action" onClick={onClick}>{action}</button></div>;
-}
-
-function Collections({
-  collections,
-  hosted,
-  cloudConfigured,
-  mirrors,
-  mirrorTarget,
-  authorityConflicts,
-  busy,
-  copiedCollectionPath,
-  onAdd,
-  onCancelCopy,
-  onCreate,
-  onRegisterCopy,
-  onMirrorTargetHandled,
-  onAct,
-  onNotice
-}: {
-  collections: CollectionSummary[];
-  hosted: HostedControlSnapshot;
-  cloudConfigured: boolean;
-  mirrors: DesktopMirrorSummary[];
-  mirrorTarget: string | null;
-  authorityConflicts: AuthorityConflict[];
-  busy: boolean;
-  copiedCollectionPath: string | null;
-  onAdd(): void;
-  onCancelCopy(): void;
-  onCreate(): void;
-  onRegisterCopy(): void;
-  onMirrorTargetHandled(): void;
-  onAct(action: () => Promise<void>): Promise<void>;
-  onNotice(value: string): void;
-}) {
-  return (
-    <section className="collection-section">
-      <SectionHeading title="Collections" note="Authority and local copies are shown separately.">
-        <button className="button secondary" disabled={busy} onClick={onAdd}>Add existing</button>
-        <button className="button primary" disabled={busy} onClick={onCreate}>Create collection</button>
-      </SectionHeading>
-      {copiedCollectionPath && (
-        <section className="copy-registration" aria-labelledby="copy-title">
-          <div>
-            <p className="eyebrow">Copied collection</p>
-            <h2 id="copy-title">Register this copy independently?</h2>
-            <p>This folder has the same Connect identity as a registered collection. Continuing writes a new identity only to the selected copy’s <code>mdbase.yaml</code>. The original is not changed.</p>
-            <code>{copiedCollectionPath}</code>
-            <small>Apps will treat the copy as a separate collection with its own links and access approvals.</small>
-          </div>
-          <div className="copy-registration-actions">
-            <button className="button secondary" disabled={busy} onClick={onCancelCopy}>Cancel</button>
-            <button className="button primary" disabled={busy} onClick={onRegisterCopy}>Register copy</button>
-          </div>
-        </section>
-      )}
-      {authorityConflicts.map((conflict) => {
-        const selectedFolder = collections.find(
-          (collection) => collection.id === conflict.collection_id
-        )?.path ?? conflict.display_name;
-        return <section className="copy-registration" aria-labelledby={`authority-${conflict.collection_id}`} key={conflict.collection_id}>
-          <div>
-            <p className="eyebrow">Collection identity conflict</p>
-            <h2 id={`authority-${conflict.collection_id}`}>Choose which copy of {conflict.display_name} to use.</h2>
-            <p>The selected folder and an existing connected copy share the same collection ID.</p>
-            <dl className="identity-conflict-details">
-              <div><dt>Selected folder</dt><dd><code title={selectedFolder}>{selectedFolder}</code></dd></div>
-              <div><dt>Currently active through</dt><dd>{conflict.active_connector_name}</dd></div>
-            </dl>
-            <small>Using the selected folder moves authority here and revokes application access through {conflict.active_connector_name}. Keeping both writes a new ID only to the selected folder’s <code>mdbase.yaml</code>.</small>
-          </div>
-          <div className="copy-registration-actions">
-            <button className="button secondary" disabled={busy} onClick={() => void onAct(async () => {
-              const independent = await window.mdbaseConnect.makeCollectionIndependent(conflict.collection_id);
-              onNotice(`${independent.display_name} now has an independent collection identity.`);
-            })}>Keep both copies</button>
-            <button className="button primary" disabled={busy} onClick={() => void onAct(async () => {
-              if (!window.confirm(`Use ${selectedFolder} as the authority for ${conflict.display_name}? Existing application access through ${conflict.active_connector_name} will be revoked.`)) return;
-              await window.mdbaseConnect.takeCollectionAuthority(conflict.collection_id);
-              onNotice(`${conflict.display_name} now uses ${selectedFolder} as its authoritative folder.`);
-            })}>Use selected folder</button>
-          </div>
-        </section>;
-      })}
-      <div className="collection-authority-group">
-        <SectionHeading title="On this computer" note="These folders are authoritative here." count={collections.length} />
-        {collections.length === 0 ? (
-          <Empty title="No computer-owned collections" text="Add a folder with an existing mdbase.yaml, or create one here." />
-        ) : (
-          <div className="collection-list">
-            {collections.map((collection) => (
-              <CollectionRow
-                key={collection.id}
-                collection={collection}
-                cloudConfigured={cloudConfigured}
-                busy={busy}
-                onAct={onAct}
-                onNotice={onNotice}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-      <div className="collection-authority-group">
-        <SectionHeading
-          title="Hosted by mdbase"
-          note={!cloudConfigured
-            ? "Connect this computer to manage hosted collections."
-            : hosted.online
-              ? "Available to approved apps without this computer."
-              : "Hosted controls are offline; last known state is shown."}
-          count={hosted.hosted_collections.length}
-        />
-        {hosted.hosted_collections.length === 0 ? (
-          <Empty title="No hosted collections" text="Create one to keep its authority online, with an optional folder mirror here." />
-        ) : (
-          <div className="collection-list">
-            {hosted.hosted_collections.map((collection) => (
-              <HostedCollectionRow
-                key={collection.id}
-                collection={collection}
-                mirrors={mirrors}
-                openMirror={mirrorTarget === collection.id}
-                busy={busy}
-                onTargetHandled={onMirrorTargetHandled}
-                onAct={onAct}
-                onNotice={onNotice}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-    </section>
-  );
-}
-
-function CollectionRow({ collection, cloudConfigured, busy, onAct, onNotice }: {
-  collection: CollectionSummary;
-  cloudConfigured: boolean;
-  busy: boolean;
-  onAct(action: () => Promise<void>): Promise<void>;
-  onNotice(value: string): void;
-}) {
-  const [editing, setEditing] = useState(false);
-  const [name, setName] = useState(collection.display_name);
-  const [description, setDescription] = useState(collection.description ?? "");
-  useEffect(() => {
-    if (!editing) {
-      setName(collection.display_name);
-      setDescription(collection.description ?? "");
-    }
-  }, [collection.description, collection.display_name, editing]);
-
-  const changed = name.trim() !== collection.display_name
-    || description.trim() !== (collection.description ?? "");
-  return (
-    <article className={`collection-card ${editing ? "editing" : ""}`}>
-      <div className="collection-summary">
-        <div className="collection-copy">
-          <div className="collection-title-row"><h3>{collection.display_name}</h3><span className="version">v{collection.spec_version}</span></div>
-          {collection.description && <p>{collection.description}</p>}
-          <button className="path" title={collection.path} onClick={() => void window.mdbaseConnect.openPath(collection.path)}>{collection.path}</button>
-        </div>
-        <div className="collection-status"><StatusDot state={collection.enabled ? "connected" : "idle"} />{collection.enabled ? "Available" : "Disabled"}</div>
-        <div className="row-actions">
-          <button
-            className="quiet-action"
-            disabled={busy}
-            onClick={() => void window.mdbaseConnect.openEditor(collection.id)}
-          >
-            Open in editor <span aria-hidden="true">↗</span>
-          </button>
-          <button className="quiet-action" disabled={busy} aria-expanded={editing} onClick={() => setEditing((value) => !value)}>{editing ? "Close" : "Details"}</button>
-          <button className="quiet-action" disabled={busy} onClick={() => void onAct(async () => { await window.mdbaseConnect.setCollectionEnabled(collection.id, !collection.enabled); onNotice(collection.enabled ? `${collection.display_name} is no longer available to remote applications.` : `${collection.display_name} is available again.`); })}>{collection.enabled ? "Disable" : "Enable"}</button>
-        </div>
-      </div>
-      {editing && <div className="collection-editor">
-        <form className="collection-editor-form" onSubmit={(event) => { event.preventDefault(); void onAct(async () => { const updated = await window.mdbaseConnect.updateCollectionMetadata({ collectionId: collection.id, name, description }); setEditing(false); onNotice(`${updated.display_name} details were saved to mdbase.yaml.`); }); }}>
-          <section className="collection-editor-section">
-            <div><strong>Details</strong><small>Name and description are stored in mdbase.yaml.</small></div>
-            <div className="collection-fields">
-              <label><span>Name</span><input value={name} maxLength={100} required onChange={(event) => setName(event.target.value)} /></label>
-              <label><span>Description</span><textarea value={description} maxLength={500} rows={2} placeholder="Optional" onChange={(event) => setDescription(event.target.value)} /></label>
-            </div>
-          </section>
-          <section className="collection-editor-section">
-            <div><strong>Configuration</strong><small>Inspect the source file or check the collection structure.</small></div>
-            <div className="collection-config-actions">
-              <button type="button" className="quiet-action" disabled={busy} onClick={() => void window.mdbaseConnect.openCollectionConfig(collection.id)}>Open mdbase.yaml</button>
-              <button type="button" className="quiet-action" disabled={busy} onClick={() => void onAct(async () => { await window.mdbaseConnect.validateCollection(collection.id); onNotice(`${collection.display_name} passed collection validation.`); })}>Validate collection</button>
-            </div>
-          </section>
-          <footer className="request-footer collection-editor-footer">
-            <p>Saving changes updates collection metadata without moving or rewriting records.</p>
-            <button className="button primary" disabled={busy || !changed || !name.trim()}>Save details</button>
-          </footer>
-        </form>
-        <section className="collection-editor-section">
-          <div>
-            <strong>Authority</strong>
-            <small>Move the source of truth online while keeping this folder as a two-way mirror.</small>
-          </div>
-          <div className="collection-config-actions">
-            <button
-              type="button"
-              className="button secondary"
-              disabled={busy || !cloudConfigured || !collection.enabled}
-              onClick={() => {
-                if (!window.confirm(
-                  `Move ${collection.display_name} authority online? `
-                  + "The collection will be uploaded directly to the provider, existing local app grants will be revoked, "
-                  + "and this folder will become a two-way mirror."
-                )) return;
-                void onAct(async () => {
-                  const result = await window.mdbaseConnect.transferCollectionAuthority(collection.id);
-                  setEditing(false);
-                  onNotice(
-                    `${collection.display_name} is now online at authority epoch `
-                    + `${result.transfer.authority_epoch}; this folder is its two-way mirror.`
-                  );
-                });
-              }}
-            >
-              Move authority online
-            </button>
-            {!cloudConfigured && <small>Connect this computer to an account first.</small>}
-          </div>
-        </section>
-        <div className="collection-danger-row">
-          <small>Removing this collection from mdbase connect never deletes its files.</small>
-          <button className="quiet-action danger" disabled={busy} onClick={() => { if (window.confirm(`Remove ${collection.display_name} from mdbase connect? Its files will not be deleted.`)) void onAct(async () => { await window.mdbaseConnect.removeCollection(collection.id); onNotice(`${collection.display_name} was removed.`); }); }}>Remove from mdbase connect</button>
-        </div>
-      </div>}
-    </article>
-  );
-}
-
-function HostedCollectionRow({
-  collection,
-  mirrors,
-  openMirror,
-  busy,
-  onTargetHandled,
-  onAct,
-  onNotice
-}: {
-  collection: HostedCollectionSummary;
-  mirrors: DesktopMirrorSummary[];
-  openMirror: boolean;
-  busy: boolean;
-  onTargetHandled(): void;
-  onAct(action: () => Promise<void>): Promise<void>;
-  onNotice(value: string): void;
-}) {
-  const [editing, setEditing] = useState(openMirror);
-  const [name, setName] = useState(collection.display_name);
-  const [path, setPath] = useState("");
-  const [mode, setMode] = useState<"read_only" | "read_write">("read_write");
-  const [promotionStarting, setPromotionStarting] = useState(false);
-  const mirror = mirrors.find((candidate) => candidate.collection_id === collection.id);
-  const activeReplicas = collection.replicas.filter((replica) => replica.revoked_at === null);
-  const editorCollectionId = collection.authority_state === "active"
-    ? collection.id
-    : collection.authority_state === "transferred"
-      ? collection.transferred_collection_id
-      : null;
-
-  useEffect(() => {
-    if (openMirror) {
-      setEditing(true);
-      onTargetHandled();
-    }
-  }, [onTargetHandled, openMirror]);
-  useEffect(() => {
-    if (!editing) setName(collection.display_name);
-  }, [collection.display_name, editing]);
-
-  async function chooseMirrorFolder() {
-    const selected = await window.mdbaseConnect.chooseMirrorFolder();
-    if (selected) setPath(selected);
-  }
-
-  const state = mirrorState(mirror);
-  const promotion = authorityPromotionState(collection, mirror, promotionStarting);
-  return (
-    <article className={`collection-card hosted-collection ${editing ? "editing" : ""}`}>
-      <div className="collection-summary">
-        <div className="collection-copy">
-          <div className="collection-title-row">
-            <h3>{collection.display_name}</h3>
-            <span className="version">v{collection.spec_version}</span>
-          </div>
-          <span className="authority-label">
-            {collection.authority_state === "transferred"
-              ? "Retired hosted copy · authority moved"
-              : collection.authority_state === "transferring"
-                ? "Authority transfer in progress"
-                : `Hosted authority · ${activeReplicas.length} ${plural(activeReplicas.length, "mirror", "mirrors")}`}
-          </span>
-        </div>
-        <div className="collection-status">
-          <StatusDot state={collection.authority_state === "active" ? "connected" : "idle"} />
-          {collection.authority_state === "active"
-            ? "Available"
-            : collection.authority_state === "transferred"
-              ? "Moved"
-              : "Moving"}
-        </div>
-        <div className="row-actions">
-          {editorCollectionId && <button
-            className="quiet-action"
-            disabled={busy}
-            onClick={() => void window.mdbaseConnect.openEditor(editorCollectionId)}
-          >
-            Open in editor <span aria-hidden="true">↗</span>
-          </button>}
-          <button className="quiet-action" disabled={busy} aria-expanded={editing} onClick={() => setEditing((value) => !value)}>{editing ? "Close" : mirror ? "Mirror" : "Details"}</button>
-        </div>
-      </div>
-      {editing && <div className="collection-editor hosted-editor">
-        <form className="collection-editor-form" onSubmit={(event) => {
-          event.preventDefault();
-          void onAct(async () => {
-            await window.mdbaseConnect.renameHostedCollection({ collectionId: collection.id, name });
-            onNotice(`${name.trim()} was renamed.`);
-          });
-        }}>
-          <section className="collection-editor-section">
-            <div><strong>Details</strong><small>The name is stored with the hosted authority.</small></div>
-            <div className="collection-fields">
-              <label><span>Name</span><input value={name} maxLength={200} required onChange={(event) => setName(event.target.value)} /></label>
-              <button className="button secondary" disabled={busy || !name.trim() || name.trim() === collection.display_name}>Save name</button>
-            </div>
-          </section>
-        </form>
-        {collection.authority_state !== "transferred" && <section className="collection-editor-section mirror-section">
-          <div>
-            <strong>Mirror on this computer</strong>
-            <small>A mirror is a local copy. The hosted collection remains authoritative.</small>
-          </div>
-          {mirror ? (
-            <div className="mirror-control">
-              <div className="mirror-state-row">
-                <StatusDot state={state.dot} />
-                <div><strong>{state.label}</strong><button className="path" title={mirror.path} onClick={() => void window.mdbaseConnect.openMirror(mirror.replica_id)}>{mirror.path}</button></div>
-                <code>{mirror.mode === "read_write" ? "two-way" : "receive-only"}</code>
-              </div>
-              {mirror.progress && <small>{mirror.progress.phase === "uploading" ? "Uploading" : "Applying"} {mirror.progress.completed}{mirror.progress.total === null ? "" : ` of ${mirror.progress.total}`} changes…</small>}
-              {mirror.error && <div className="message error-message compact-message">{mirror.error}</div>}
-              {mirror.conflicts.length > 0 && (
-                <div className="mirror-conflicts">
-                  {mirror.conflicts.map((conflict) => <div key={conflict.record_id}>
-                    <div><strong>{conflict.path ?? conflict.record_id}</strong><small>{conflict.message}</small></div>
-                    <div className="row-actions">
-                      <button className="quiet-action" disabled={busy} onClick={() => void onAct(async () => {
-                        await window.mdbaseConnect.resolveMirrorConflict({ replicaId: mirror.replica_id, recordId: conflict.record_id, resolution: "local" });
-                        onNotice("The local version was kept and synchronized.");
-                      })}>Keep local</button>
-                      <button className="quiet-action" disabled={busy} onClick={() => void onAct(async () => {
-                        await window.mdbaseConnect.resolveMirrorConflict({ replicaId: mirror.replica_id, recordId: conflict.record_id, resolution: "remote" });
-                        onNotice("The hosted version was applied.");
-                      })}>Use hosted</button>
-                    </div>
-                  </div>)}
-                </div>
-              )}
-              {mirror.local_issues.length > 0 && (
-                <div className="mirror-conflicts">
-                  {mirror.local_issues.map((issue) => <div key={issue.path}>
-                    <div>
-                      <strong>{issue.path}</strong>
-                      <small>{issue.message} Other valid Markdown continues to synchronize.</small>
-                    </div>
-                  </div>)}
-                </div>
-              )}
-              <div className="mirror-actions">
-                <button className="quiet-action" disabled={busy || mirror.syncing} onClick={() => void onAct(async () => {
-                  await window.mdbaseConnect.syncMirror(mirror.replica_id);
-                  onNotice(`${collection.display_name} is synchronized.`);
-                })}>{mirror.syncing ? "Synchronizing…" : "Sync now"}</button>
-                <button className="quiet-action" disabled={busy} onClick={() => void window.mdbaseConnect.openMirror(mirror.replica_id)}>Open folder</button>
-                <button className="quiet-action danger" disabled={busy} onClick={() => {
-                  if (!window.confirm(`Stop mirroring ${collection.display_name} on this computer? The folder and its files will remain.`)) return;
-                  void onAct(async () => {
-                    await window.mdbaseConnect.disconnectMirror(mirror.replica_id);
-                    onNotice(`The mirror was disconnected. Files remain at ${mirror.path}.`);
-                  });
-                }}>Stop mirror</button>
-              </div>
-            </div>
-          ) : (
-            <div className="mirror-setup">
-              <label><span>Folder</span><button type="button" className="folder-picker" onClick={() => void chooseMirrorFolder()}>{path || "Choose a folder…"}</button></label>
-              <label><span>Synchronization</span><select value={mode} onChange={(event) => setMode(event.target.value as "read_only" | "read_write")}><option value="read_write">Two-way</option><option value="read_only">Receive-only</option></select></label>
-              <small>Two-way mirrors upload local edits. Receive-only mirrors replace their local view with hosted changes.</small>
-              <button className="button primary" disabled={busy || !path} onClick={() => void onAct(async () => {
-                await window.mdbaseConnect.connectMirror({ collectionId: collection.id, path, mode });
-                setPath("");
-                onNotice(`${collection.display_name} is now mirrored on this computer.`);
-              })}>Start mirror</button>
-            </div>
-          )}
-        </section>}
-        {collection.authority_state === "transferred" ? (
-          <section className="collection-editor-section">
-            <div>
-              <strong>Authority</strong>
-              <small>This hosted copy is retained for recovery but no longer accepts changes.</small>
-            </div>
-            <div className="authority-transfer-control">
-              <div>
-                <strong>Source of truth moved to this computer</strong>
-                <small>The collection now appears under On this computer. Applications need fresh access to the computer-owned collection.</small>
-              </div>
-            </div>
-          </section>
-        ) : mirror && (
-          <section className="collection-editor-section">
-            <div>
-              <strong>Authority</strong>
-              <small>Make this synchronized folder the source of truth.</small>
-            </div>
-            <div className="authority-transfer-control">
-              <div>
-                <strong>{promotion.title}</strong>
-                <small>{promotion.detail}</small>
-              </div>
-              <button
-                type="button"
-                className="button secondary"
-                disabled={busy || !promotion.enabled}
-                onClick={() => {
-                  if (
-                    !mirror.promotion_pending
-                    && !window.confirm(
-                      `Move ${collection.display_name} authority to this computer? `
-                      + `${mirror.path} will become the source of truth. Hosted writes will stop, `
-                      + "and existing application access and other mirrors will be revoked. "
-                      + "You will confirm this change in your browser."
-                    )
-                  ) return;
-                  setPromotionStarting(true);
-                  void onAct(async () => {
-                    onNotice(
-                      mirror.promotion_pending
-                        ? "Resuming the authority transfer. Keep mdbase connect open."
-                        : "Confirm the authority transfer in your browser, then return to mdbase connect."
-                    );
-                    try {
-                      const result = await window.mdbaseConnect.promoteMirrorAuthority(
-                        mirror.replica_id
-                      );
-                      setEditing(false);
-                      onNotice(
-                        `${collection.display_name} now uses this computer as its authority at epoch `
-                        + `${result.authority_epoch}. Applications need fresh access.`
-                      );
-                    } finally {
-                      setPromotionStarting(false);
-                    }
-                  });
-                }}
-              >
-                {promotion.button}
-              </button>
-            </div>
-          </section>
-        )}
-        {activeReplicas.some((replica) => replica.id !== mirror?.replica_id) && (
-          <section className="collection-editor-section">
-            <div><strong>Other mirrors</strong><small>Mirrors connected from another installation or through the command line.</small></div>
-            <div className="replica-list">{activeReplicas.filter((replica) => replica.id !== mirror?.replica_id).map((replica) => (
-              <div key={replica.id}><span>{replica.name}</span><code>{replica.mode === "read_write" ? "two-way" : "receive-only"}</code><small>{replica.sync_status?.last_seen_at ? `Seen ${relativeTime(replica.sync_status.last_seen_at)}` : "Not synchronized yet"}</small><button className="quiet-action danger" disabled={busy} onClick={() => {
-                if (!window.confirm(`Revoke ${replica.name}? Its local files will remain, but it will no longer synchronize.`)) return;
-                void onAct(async () => {
-                  await window.mdbaseConnect.revokeHostedReplica(replica.id);
-                  onNotice(`${replica.name} was revoked.`);
-                });
-              }}>Revoke</button></div>
-            ))}</div>
-          </section>
-        )}
-        <div className="collection-danger-row">
-          <small>Deleting a hosted collection permanently removes its hosted records. Local mirror files remain.</small>
-          <button className="quiet-action danger" disabled={busy} onClick={() => {
-            if (!window.confirm(`Permanently delete the hosted collection ${collection.display_name}? This cannot be undone.`)) return;
-            void onAct(async () => {
-              if (mirror) await window.mdbaseConnect.disconnectMirror(mirror.replica_id);
-              await window.mdbaseConnect.deleteHostedCollection(collection.id);
-              onNotice(`${collection.display_name} was deleted. Any local mirror files remain.`);
-            });
-          }}>Delete hosted collection</button>
-        </div>
-      </div>}
-    </article>
-  );
 }
 
 function Access({ cloud, access, collections, hostedCollections, canCreateHosted, busy, onAct, onNotice }: {
@@ -1397,318 +908,6 @@ function ThemeSelect() {
     setPreference(next);
     saveThemePreference(next);
   }}><option value="system">System</option><option value="light">Light</option><option value="dark">Dark</option></select>;
-}
-
-function PairingPanel() {
-  const [serverUrl, setServerUrl] = useState("https://connect.mdbase.dev");
-  const [connectorName, setConnectorName] = useState("This computer");
-  const [pairing, setPairing] = useState<{ pairingId: string; verificationUri: string } | null>(null);
-  const [pairError, setPairError] = useState("");
-  const [starting, setStarting] = useState(false);
-  const [completing, setCompleting] = useState(false);
-
-  useEffect(() => {
-    if (!pairing) return;
-    const timer = window.setInterval(async () => {
-      try {
-        const result = await window.mdbaseConnect.pairingStatus(pairing.pairingId);
-        if (result.status === "paired") {
-          window.clearInterval(timer);
-          setCompleting(true);
-        }
-      } catch (error) {
-        setPairError(message(error));
-        window.clearInterval(timer);
-      }
-    }, 2_000);
-    return () => window.clearInterval(timer);
-  }, [pairing]);
-
-  async function begin(event: React.FormEvent) {
-    event.preventDefault();
-    setStarting(true);
-    setPairError("");
-    try {
-      const result = await window.mdbaseConnect.beginPairing({ serverUrl, connectorName });
-      setPairing(result);
-    } catch (error) {
-      setPairError(message(error));
-    } finally {
-      setStarting(false);
-    }
-  }
-
-  return (
-    <section className="pairing-panel">
-      <div className="pairing-intro"><p className="eyebrow">Portal connection</p><h2>{pairing ? "Finish in your browser." : "Connect this computer."}</h2><p>{pairing ? "Sign in and approve the computer. This window will update automatically, and no token needs to be copied." : "A portal supplies identity and routing so authorized websites can find this connector. Collection paths remain local."}</p></div>
-      {pairError && <div className="message error-message">{pairError}</div>}
-      {pairing ? (
-        <div className="pairing-wait" role="status" aria-live="polite">
-          <StatusDot state="connecting" />
-          <div>
-            <strong>{completing ? "Computer approved. Connecting securely…" : "Waiting for browser approval"}</strong>
-            {completing ? <small>mdbase connect is restarting with the new secure connection.</small> : <code>{pairing.verificationUri}</code>}
-          </div>
-          {!completing && <button className="quiet-action" onClick={() => setPairing(null)}>Start again</button>}
-        </div>
-      ) : (
-        <form className="pairing-form" onSubmit={(event) => void begin(event)}>
-          <label><span>Server</span><input type="url" value={serverUrl} onChange={(event) => setServerUrl(event.target.value)} /></label>
-          <label><span>Computer name</span><input value={connectorName} onChange={(event) => setConnectorName(event.target.value)} /></label>
-          <button className="button primary" disabled={starting || !serverUrl.trim() || !connectorName.trim()}>{starting ? "Opening browser…" : "Continue in browser"}</button>
-        </form>
-      )}
-    </section>
-  );
-}
-
-function NavButton({ route, current, label, count, attention, onSelect }: { route: Route; current: Route; label: string; count?: number; attention?: number; onSelect(route: Route): void }) {
-  return <button className={`view-tab ${current === route ? "active" : ""}`} aria-current={current === route ? "page" : undefined} onClick={() => onSelect(route)}><span>{label}</span>{attention ? <b className="view-tab-count attention">{attention}</b> : count !== undefined ? <b className="view-tab-count">{count}</b> : null}</button>;
-}
-
-function SectionHeading({ title, note, count, children }: { title: string; note: string; count?: number; children?: React.ReactNode }) {
-  return <div className="section-heading"><div><h2>{title}</h2><p>{note}</p></div><div className="heading-actions">{count !== undefined && <span className="count">{count}</span>}{children}</div></div>;
-}
-
-function Empty({ title, text, action, onAction }: { title: string; text: string; action?: string; onAction?(): void }) {
-  return <div className="empty-state"><div className="empty-folder" aria-hidden="true"><span /></div><h3>{title}</h3><p>{text}</p>{action && onAction && <button className="text-action" onClick={onAction}>{action}</button>}</div>;
-}
-
-function StatusDot({ state }: { state: ConnectionDotState }) {
-  return <span className={`status-dot ${state}`} aria-hidden="true" />;
-}
-
-function Brand() {
-  return <div className="product-brand"><MdbaseMark /><strong>mdbase</strong><span className="product-brand-label">connect</span></div>;
-}
-
-function MdbaseMark() {
-  return <svg className="product-brand-mark" viewBox={MDBASE_MARK_VIEW_BOX} aria-hidden="true" focusable="false">
-    <g className="product-brand-mark-ink">
-      {mdbaseMarkInkRects.map((rect) => <rect key={`${rect.x}-${rect.y}`} {...rect} />)}
-    </g>
-    <rect className="product-brand-mark-accent" {...mdbaseMarkAccentRect} />
-  </svg>;
-}
-
-function SettingSwitch({ className, label, description, checked, disabled, stateLabel, onChange }: {
-  className: "pause-control" | "setting-toggle";
-  label: string;
-  description: string;
-  checked: boolean;
-  disabled: boolean;
-  stateLabel: string;
-  onChange(checked: boolean): void;
-}) {
-  return (
-    <label className={`${className} ${disabled ? "disabled" : ""}`}>
-      <span className="toggle-copy"><strong>{label}</strong><small>{description}</small></span>
-      <span className="toggle-action">
-        <span className="toggle-state" aria-hidden="true">{stateLabel}</span>
-        <input
-          type="checkbox"
-          role="switch"
-          aria-label={label}
-          checked={checked}
-          disabled={disabled}
-          onChange={(event) => onChange(event.target.checked)}
-        />
-      </span>
-    </label>
-  );
-}
-
-function neededProvisions(
-  requirements: ApplicationRequirements,
-  provisions: ApplicationProvisions | undefined,
-  collection: Pick<AuthorizationCollection, "contracts">
-): TypePackProvision[] {
-  const missing = requirements.contracts.filter((requirement) => !hasContract(collection.contracts, requirement));
-  return (provisions?.type_packs ?? []).filter((provision) =>
-    provision.provides.some((provided) => missing.some((requirement) => sameContract(provided, requirement)))
-  );
-}
-
-const MDBASE_03_OPERATIONS = new Set([
-  "query",
-  "list_views",
-  "execute_view",
-  "read_type",
-  "create_type",
-  "update_type"
-]);
-
-function hostedCollectionCompatible(
-  request: PendingAuthorization,
-  collection: HostedCollectionSummary
-): boolean {
-  if (
-    request.requested_operations.some((operation) => MDBASE_03_OPERATIONS.has(operation))
-    && !/^0\.3(?:\.|$)/.test(collection.spec_version)
-  ) {
-    return false;
-  }
-  return request.requirements.contracts.every((required) =>
-    hasContract(collection.contracts, required)
-    || request.provisions.type_packs.some((provision) =>
-      provision.provides.some((provided) => sameContract(provided, required))
-    )
-  );
-}
-
-function mirrorState(mirror: DesktopMirrorSummary | undefined): {
-  dot: ConnectionDotState;
-  label: string;
-} {
-  if (!mirror) return { dot: "idle", label: "Not mirrored" };
-  if (mirror.promotion) {
-    return {
-      dot: "connecting",
-      label: authorityPromotionPhaseLabel(mirror.promotion.phase)
-    };
-  }
-  if (mirror.syncing) return { dot: "connecting", label: "Synchronizing" };
-  if (mirror.error || mirror.state === "offline") return { dot: "danger", label: "Mirror needs attention" };
-  if (mirror.conflicts.length > 0) return { dot: "paused", label: "Conflicts need a decision" };
-  if (mirror.local_issues.length > 0 || mirror.state === "attention") return { dot: "paused", label: "Local files need attention" };
-  if (mirror.state === "changes_waiting" || mirror.pending > 0) return { dot: "connecting", label: "Changes waiting" };
-  if (mirror.state === "not_initialized") return { dot: "idle", label: "Not synchronized yet" };
-  return { dot: "connected", label: "Up to date" };
-}
-
-function authorityPromotionState(
-  collection: HostedCollectionSummary,
-  mirror: DesktopMirrorSummary | undefined,
-  starting: boolean
-): { enabled: boolean; title: string; detail: string; button: string } {
-  if (!mirror) {
-    return {
-      enabled: false,
-      title: "A two-way mirror is required",
-      detail: "Create and synchronize a two-way mirror on this computer first.",
-      button: "Move authority to this computer"
-    };
-  }
-  if (starting && !mirror.promotion) {
-    return {
-      enabled: false,
-      title: "Preparing authority transfer",
-      detail: "Checking the folder before opening browser confirmation.",
-      button: "Preparing transfer…"
-    };
-  }
-  if (mirror.promotion) {
-    const label = authorityPromotionPhaseLabel(mirror.promotion.phase);
-    return {
-      enabled: false,
-      title: label,
-      detail: mirror.promotion.phase === "awaiting_approval"
-        ? "Approve the move in your browser. Hosted writes continue until approval."
-        : "Keep mdbase connect open while the handoff finishes.",
-      button: `${label}…`
-    };
-  }
-  if (mirror.promotion_pending) {
-    return {
-      enabled: true,
-      title: "Authority transfer ready to resume",
-      detail: "The folder was already verified. Finish activating it as the source of truth.",
-      button: "Resume authority move"
-    };
-  }
-  if (mirror.mode !== "read_write") {
-    return {
-      enabled: false,
-      title: "A two-way mirror is required",
-      detail: "Receive-only folders cannot become authoritative. Recreate this mirror as two-way.",
-      button: "Move authority to this computer"
-    };
-  }
-  if (collection.authority_state !== "active") {
-    return {
-      enabled: false,
-      title: "Authority transfer already in progress",
-      detail: "Return to the browser confirmation or wait for the current request to expire.",
-      button: "Move authority to this computer"
-    };
-  }
-  if (mirror.syncing) {
-    return {
-      enabled: false,
-      title: "Mirror is synchronizing",
-      detail: "Authority can move after the current synchronization finishes.",
-      button: "Move authority to this computer"
-    };
-  }
-  if (
-    mirror.error
-    || mirror.state === "offline"
-    || mirror.conflicts.length > 0
-    || mirror.local_issues.length > 0
-    || mirror.state === "attention"
-  ) {
-    return {
-      enabled: false,
-      title: "Mirror needs attention",
-      detail: "Resolve mirror errors and file conflicts before moving authority.",
-      button: "Move authority to this computer"
-    };
-  }
-  if (mirror.state !== "up_to_date" || mirror.pending > 0) {
-    return {
-      enabled: false,
-      title: "Synchronize this mirror first",
-      detail: "The folder must match the hosted collection before it can become authoritative.",
-      button: "Move authority to this computer"
-    };
-  }
-  return {
-    enabled: true,
-    title: "Move authority to this computer",
-    detail: "Browser confirmation will pause hosted writes, verify this folder, and revoke existing application access.",
-    button: "Move authority to this computer"
-  };
-}
-
-function authorityPromotionPhaseLabel(
-  phase: NonNullable<DesktopMirrorSummary["promotion"]>["phase"]
-): string {
-  if (phase === "synchronizing") return "Synchronizing mirror";
-  if (phase === "awaiting_approval") return "Waiting for browser approval";
-  if (phase === "verifying") return "Verifying exact folder contents";
-  if (phase === "registering") return "Registering local authority";
-  if (phase === "registered" || phase === "activating") return "Activating local authority";
-  if (phase === "resuming") return "Resuming authority transfer";
-  return "Finishing authority transfer";
-}
-
-function hasContract(contracts: ContractRequirement[], required: ContractRequirement) { return contracts.some((contract) => sameContract(contract, required)); }
-function sameContract(left: ContractRequirement, right: ContractRequirement) { return left.id === right.id && left.version === right.version; }
-function provisionNames(provisions: TypePackProvision[]) {
-  return provisions
-    .map((provision) => provision.manifest.name ?? provision.manifest.id)
-    .join(" and ");
-}
-
-function scopeDescription(contracts: ContractRequirement[]): string {
-  const names = contracts.map((contract) => `${contract.id} v${contract.version}`);
-  return `Records matching ${names.join(" and ")} only`;
-}
-
-function host(value: string) { try { return new URL(value).host; } catch { return value; } }
-function initials(value: string) { return value.split(/\s+/).map((part) => part[0]).join("").slice(0, 2).toUpperCase(); }
-function message(error: unknown): string { return error instanceof Error ? error.message : String(error); }
-function plural(count: number, singular: string, pluralValue: string) { return count === 1 ? singular : pluralValue; }
-function relativeTime(value: string) {
-  const date = new Date(value.includes("T") ? value : `${value.replace(" ", "T")}Z`);
-  const seconds = Math.round((date.getTime() - Date.now()) / 1_000);
-  const formatter = new Intl.RelativeTimeFormat(undefined, { numeric: "auto" });
-  if (Math.abs(seconds) < 60) return formatter.format(seconds, "second");
-  const minutes = Math.round(seconds / 60);
-  if (Math.abs(minutes) < 60) return formatter.format(minutes, "minute");
-  const hours = Math.round(minutes / 60);
-  if (Math.abs(hours) < 24) return formatter.format(hours, "hour");
-  return formatter.format(Math.round(hours / 24), "day");
 }
 
 createRoot(document.getElementById("root")!).render(<React.StrictMode><App /></React.StrictMode>);
